requirements.txt
colored_traceback==0.3.0
h5py==3.7.0
matplotlib==3.5.3
nltk==3.7
numpy==1.23.2
opencv_python==4.6.0.66
Pillow==9.2.0
pycocoevalcap==1.2
scikit_image==0.19.3
skimage==0.0
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.64.0
